modified shared document entity

+++ src/services/api/http/interceptors/connection.js
@@ -1,8 +1,10 @@
 export default (instance) => {
   instance.interceptors.response.use(null, err => {
     if (err.response && err.response.status === 401) {
-      window.localStorage.clear('authToken')
-      window.location.href = '/login'
+      if (window.localStorage.getItem('authToken') !== null) {
+        window.localStorage.clear('authToken')
+        window.location.href = '/login'
+      }
     }
     return Promise.reject(err)
   })
